Inline win32 dispatch into atom
